backend/requirements.txt
sentence-transformers
langchain_community
unstructured
faiss-cpu
Together
torch
numpy
